+++ projects/ontimize-web-ngx/src/lib/components/input/o-form-service-component.class.ts
@@ -5,7 +5,6 @@ import { BooleanInputConverter } from '../../decorators/input-converter';
 import { ServiceResponse } from '../../interfaces/service-response.interface';
 import { OErrorDialogManager } from '../../services/o-error-dialog-manager.service';
 import { OntimizeService } from '../../services/ontimize/ontimize.service';
-import { OntimizeQueryArgumentsAdapter } from '../../services/query-arguments/ontimize-query-arguments.adapter';
 import { OConfigureServiceArgs } from '../../types/configure-service-args.type';
 import { FormValueOptions } from '../../types/form-value-options.type';
 import { OQueryDataArgs } from '../../types/query-data-args.type';
@@ -16,6 +15,7 @@ import { Util } from '../../util/util';
 import { OContextMenuComponent } from '../contextmenu/o-context-menu.component';
 import { OFormComponent } from '../form/o-form.component';
 import { OFormDataComponent } from '../o-form-data-component.class';
+import { BaseService } from '../../services/base-service.class';
 
 export const DEFAULT_INPUTS_O_FORM_SERVICE_COMPONENT = [
   // static-data [Array<any>] : way to populate with static data. Default: no value.
@@ -112,7 +112,7 @@ export class OFormServiceComponent extends OFormDataComponent {
   protected colArray: string[] = [];
   protected visibleColArray: string[] = [];
   public descriptionColArray: string[] = [];
-  protected dataService: OntimizeService;
+  protected dataService: BaseService<ServiceResponse>;
   public loaderSubscription: Subscription;
   loading: boolean = false;
 
@@ -128,7 +128,6 @@ export class OFormServiceComponent extends OFormDataComponent {
   protected subscriptionDataLoad: Subscription = new Subscription();
   public delayLoad = 250;
   public loadingSubject = new BehaviorSubject<boolean>(false);
-  queryArgumentAdapter: any;
 
   public oContextMenu: OContextMenuComponent;
   queryArguments: any;
@@ -185,7 +184,6 @@ export class OFormServiceComponent extends OFormDataComponent {
       this.setDataArray(this.staticData);
     } else {
       this.configureService();
-      this.configureAdapter();
     }
 
     if (this.queryOnEvent !== undefined && this.queryOnEvent.subscribe !== undefined) {
@@ -247,10 +245,6 @@ export class OFormServiceComponent extends OFormDataComponent {
 
   }
 
-  public configureAdapter() {
-    this.queryArgumentAdapter = this.injector.get(OntimizeQueryArgumentsAdapter);
-  }
-
   getAttributesValuesToQuery(columns?: Array<any>) {
     const result = Util.isDefined(columns) ? columns : this.colArray;
     if (result.indexOf(this.valueColumn) === -1) {
@@ -277,9 +271,8 @@ export class OFormServiceComponent extends OFormDataComponent {
 
       this.loaderSubscription = this.load();
 
-      this.queryArguments = this.queryArgumentAdapter.parseQueryParameters(this.getQueryArguments(filter));
-
-      this.querySubscription = this.queryArgumentAdapter.request.apply(this.queryArgumentAdapter, [this.queryMethod, this.dataService, this.queryArguments])
+      this.queryArguments = this.getQueryArguments(filter);
+      this.querySubscription = this.dataService[this.queryMethod](...this.dataService.queryArgumentAdapter.parseQueryParameters(this.queryArguments))
         .subscribe((resp: ServiceResponse) => {
           if (resp.isSuccessful()) {
             this.cacheQueried = true;