+++ projects/ontimize-web-ngx/src/lib/components/table/column/cell-renderer/service/o-table-cell-renderer-service.component.ts
@@ -7,8 +7,6 @@ import { ITranslatePipeArgument, OTranslatePipe } from '../../../../../pipes/o-t
 import { DialogService } from '../../../../../services/dialog.service';
 import { OntimizeServiceProvider } from '../../../../../services/factories';
 import { OntimizeService } from '../../../../../services/ontimize/ontimize.service';
-import { BaseQueryArgument } from '../../../../../services/query-arguments/base-query-argument.adapter';
-import { OntimizeQueryArgumentsAdapter } from '../../../../../services/query-arguments/ontimize-query-arguments.adapter';
 import { OConfigureServiceArgs } from '../../../../../types/configure-service-args.type';
 import { Expression } from '../../../../../types/expression.type';
 import { Codes } from '../../../../../util/codes';
@@ -19,6 +17,7 @@ import { Util } from '../../../../../util/util';
 import { OBaseTableCellRenderer } from '../o-base-table-cell-renderer.class';
 
 import type { OColumn } from '../../o-column.class';
+import { BaseService } from '../../../../../services/base-service.class';
 export const DEFAULT_INPUTS_O_TABLE_CELL_RENDERER_SERVICE = [
   'entity',
   'service',
@@ -77,7 +76,7 @@ export class OTableCellRendererServiceComponent extends OBaseTableCellRenderer i
   public onDataLoaded: EventEmitter<any> = new EventEmitter();
   /* Internal variables */
   protected colArray: string[] = [];
-  protected dataService: any;
+  protected dataService: BaseService<ServiceResponse>;
   protected _pKeysEquiv = {};
   protected dialogService: DialogService;
 
@@ -86,17 +85,12 @@ export class OTableCellRendererServiceComponent extends OBaseTableCellRenderer i
   protected pipeArguments: ITranslatePipeArgument = {};
 
   protected subscritpions: Subscription = new Subscription();
-  queryArgumentAdapter: BaseQueryArgument;
 
   constructor(protected injector: Injector) {
     super(injector);
     this.tableColumn.type = 'service';
     this.dialogService = injector.get<DialogService>(DialogService as Type<DialogService>);
     this.setComponentPipe();
-  }
-
-  public configureAdapter() {
-    this.queryArgumentAdapter = this.injector.get(OntimizeQueryArgumentsAdapter);
   }
 
   public initialize(): void {
@@ -110,7 +104,6 @@ export class OTableCellRendererServiceComponent extends OBaseTableCellRenderer i
     const pkArray = Util.parseArray(this.parentKeys);
     this._pKeysEquiv = Util.parseParentKeysEquivalences(pkArray);
     this.configureService();
-    this.configureAdapter();
   }
 
   public ngAfterViewInit(): void {
@@ -150,8 +143,8 @@ export class OTableCellRendererServiceComponent extends OBaseTableCellRenderer i
     } else {
       filter[this.column] = cellvalue;
     }
-    const queryArguments = this.queryArgumentAdapter.parseQueryParameters(this.getQueryArguments(filter));
-    this.queryArgumentAdapter.request.apply(this.queryArgumentAdapter, [this.queryMethod, this.dataService, queryArguments])
+    const queryArguments = this.getQueryArguments(filter);
+    this.dataService[this.queryMethod](...this.dataService.queryArgumentAdapter.parseQueryParameters(queryArguments))
       .subscribe((resp: ServiceResponse) => {
         if (resp.isSuccessful()) {
           let respData;
@@ -257,8 +250,9 @@ export class OTableCellRendererServiceComponent extends OBaseTableCellRenderer i
         observer.next();
       }
 
-      const queryArguments = this.queryArgumentAdapter.parseQueryParameters(this.getQueryArguments({}));
-      this.queryArgumentAdapter.request.apply(this.queryArgumentAdapter, [this.queryMethod, this.dataService, queryArguments])
+      const queryArguments = this.getQueryArguments({});
+
+      this.dataService[this.queryMethod](...this.dataService.queryArgumentAdapter.parseQueryParameters(queryArguments))
         .subscribe((resp: ServiceResponse) => {
           if (resp.isSuccessful()) {
             let respData = [];