+++ projects/ontimize-web-ngx/src/lib/components/table/o-table.component.ts
@@ -2243,8 +2243,7 @@ export class OTableComponent extends AbstractOServiceComponent<OTableComponentSt
       if (this.asyncLoadSubscriptions[rowIndex]) {
         this.asyncLoadSubscriptions[rowIndex].unsubscribe();
       }
-      this.asyncLoadSubscriptions[rowIndex] = this.dataService[queryMethodName]
-        .apply(this.dataService, columnQueryArgs)
+      this.asyncLoadSubscriptions[rowIndex] = this.dataService[queryMethodName](...columnQueryArgs)
         .subscribe((res: ServiceResponse) => {
           if (res.isSuccessful()) {
             let data;