+++ projects/ontimize-web-ngx/src/lib/interfaces/service-response-adapter.interface.ts
@@ -1,8 +1,7 @@
 import { HttpErrorResponse, HttpResponse } from '@angular/common/http';
 
 export interface IServiceResponseAdapter<T> {
-  context: any;
+  context?: any;
   adapt(res: HttpResponse<any>): T;
-  adaptError(error: HttpErrorResponse): any
-  setContext(context: any): void;
+  adaptError?(error: HttpErrorResponse): any
 }