+++ projects/ontimize-web-ngx/src/lib/services/jsonapi/jsonapi-service-response.adapter.ts
@@ -1,8 +1,8 @@
 import { HttpErrorResponse, HttpResponse } from '@angular/common/http';
 import { Injectable } from '@angular/core';
-import { JSONAPIServiceResponse } from './jsonapi-service-response.class';
+
 import { IServiceResponseAdapter } from '../../interfaces/service-response-adapter.interface';
-import { JSONAPIContext } from '../../types/jsonapi-context.type';
+import { JSONAPIServiceResponse } from './jsonapi-service-response.class';
 
 @Injectable()
 export class JSONAPIServiceResponseAdapter implements IServiceResponseAdapter<JSONAPIServiceResponse> {
@@ -45,11 +45,4 @@ export class JSONAPIServiceResponseAdapter implements IServiceResponseAdapter<JS
     }
   }
 
-  /**
-   * Sets context
-   * @param context
-   */
-  setContext(context: JSONAPIContext) {
-    this.context = context;
-  }
 }