+++ projects/ontimize-web-ngx/src/lib/services/ontimize/ontimize-service-response.adapter.ts
@@ -2,13 +2,12 @@ import { HttpErrorResponse, HttpResponse } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 
 import { IServiceResponseAdapter } from '../../interfaces/service-response-adapter.interface';
-import { ServiceResponse } from '../../interfaces/service-response.interface';
 import { OntimizeServiceResponse } from './ontimize-service-response.class';
 
 @Injectable()
 export class OntimizeServiceResponseAdapter implements IServiceResponseAdapter<OntimizeServiceResponse> {
   context:any;
-  adapt(res: HttpResponse<any>): ServiceResponse {
+  adapt(res: HttpResponse<any>): OntimizeServiceResponse {
     return new OntimizeServiceResponse(
       res.body.code,
       res.body.data,
@@ -17,10 +16,6 @@ export class OntimizeServiceResponseAdapter implements IServiceResponseAdapter<O
       res.body.startRecordIndex,
       res.body.totalQueryRecordsNumber
     );
-  }
-
-  setContext(context:any) {
-    this.context = context;
   }
 
   adaptError(error: HttpErrorResponse) {