+++ projects/ontimize-web-ngx/src/lib/types/service-configuration.type.ts
@@ -1,0 +1,8 @@
+
+export type ServiceConfigType = {
+  [key: string]: {
+    path: string;
+    fileManagerPath?: string;
+    serviceType?: string;
+  };
+};